refactor(songs): extract feed data layer into useSongFeed hook
Move the pagination state, fetching, infinite-scroll observer, and the
local list mutations out of SongsPage into a useSongFeed(view, user)
hook — matching the codebase's hook convention (useSong, useProfile,
useDragInteraction). The page component drops from ~165 lines of mixed
data/render to a thin presentational wrapper.

Behaviour is unchanged; the view-membership filter that keeps a locally
edited song in/out of the list is now a single matchesView() helper
shared by the fetch filter and the client filter.
